# server/docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "http server api docs",
        "title": "admin api docs",
        "contact": {},
        "version": "2.0"
    },
    "host": "localhost:8080",
    "paths": {
...
        "/api/v1/auth/login": {
            "post": {
                "description": "with username and password",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "with username and password",
                "parameters": [
                    {
                        "description": "login information",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/types.LoginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/types.LoginReply"
                        }
                    }
                }
            }
        },
        "/api/v1/auth/logout": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "logout",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "logout",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/types.Result"
                        }
                    }
                }
            }
        },
        "/api/v1/auth/refreshToken": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "refresh token",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "refresh token",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/types.LoginReply"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "types.LoginItem": {
            "type": "object",
            "properties": {
                "accessToken": {
                    "description": "access token",
                    "type": "string"
                },
                "expires": {
                    "description": "expire time",
                    "type": "integer"
                },
                "tokenType": {
                    "description": "token type",
                    "type": "string"
                }
            }
        },
        "types.LoginReply": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "return code",
                    "type": "integer"
                },
                "data": {
                    "description": "return data",
                    "allOf": [
                        {
                            "$ref": "#/definitions/types.LoginItem"
                        }
                    ]
                },
                "msg": {
                    "description": "return information description",
                    "type": "string"
                }
            }
        },
        "types.LoginRequest": {
            "type": "object",
            "properties": {
                "captchaCode": {
                    "description": "验证码code",
                    "type": "string"
                },
                "captchaKey": {
                    "description": "验证码key",
                    "type": "string"
                },
                "password": {
                    "description": "密码",
                    "type": "string"
                },
                "username": {
                    "description": "账号",
                    "type": "string"
                }
            }
        },
...
        "types.Result": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "return code",
                    "type": "integer"
                },
                "data": {
                    "description": "return data"
                },
                "msg": {
                    "description": "return information description",
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type Bearer your-jwt-token to Value",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}